refactor(compare): rename trending → curated-presets throughout
Rename data/trending.json → data/curated-presets.json,
lib/trending.ts → lib/curated-presets.ts, and update
TrendingPreset → CuratedPreset / trendingPresets → curatedPresets
in all consumers.
